backend-api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
fastapi==0.112.0
h11==0.14.0
httpx==0.27.2
idna==3.7
passlib==1.7.4
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
redis==5.0.8
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5